requirements.txt
Django>=5.0,<5.1
dj-database-url
whitenoise
gunicorn
psycopg2-binary
python-decouple
